+++ lib/event/data/common/event_price_response_model.dart
@@ -7,13 +7,13 @@ part 'event_price_response_model.g.dart';
 @JsonSerializable(explicitToJson: true)
 class EventPriceResponseModel extends BaseModel {
   final int? priceID;
-  final int fee;
+  final double fee;
   final EventCurrencyModel currency;
   const EventPriceResponseModel(
       {this.priceID, required this.fee, required this.currency});
 
   EventPriceResponseModel copyWith(
-      {int? priceID, int? fee, EventCurrencyModel? currency}) {
+      {int? priceID, double? fee, EventCurrencyModel? currency}) {
     return EventPriceResponseModel(
       priceID: priceID ?? this.priceID,
       fee: fee ?? this.fee,
